[Backend] Move type Role to shared
