+++ services/backend/src/services/userService.ts
@@ -6,7 +6,9 @@ import { serviceProvider } from '../config'
 import { roles } from '../config/roles'
 import { sequelizeUser } from '../database/connection'
 import { User } from '../models/user'
-import { DetailedProgrammeRights, ExpandedUser, FormattedUser, IamAccess, Role } from '../types'
+import { Role } from '../shared/types'
+import { DetailedProgrammeRights, ExpandedUser, FormattedUser, IamAccess } from '../types'
+
 import { createLocaleComparator, getFullStudyProgrammeRights, hasFullAccessToStudentData } from '../util'
 import * as jami from '../util/jami'
 import mami from '../util/mami'
