+++ services/backend/src/types/index.ts
@@ -11,7 +11,6 @@ export { GenderCode } from './genderCode'
 export { IamAccess } from './iamAccess'
 export { Language } from './language'
 export { Phase } from './phase'
-export { Role } from './role'
 export { SemesterEnrollment } from './semesterEnrollment'
 export { StudyTrack } from './studyTrack'
 export { Unification, UnifyStatus } from './unification'
