--- services/backend/src/types/role.ts
@@ -1,3 +1,0 @@
-import { roles } from '../config/roles'
-
-export type Role = (typeof roles)[number]
