+++ services/backend/src/types/user.ts
@@ -1,9 +1,9 @@
 import { InferAttributes } from 'sequelize'
 
 import { User } from '../models/user'
+import { Role } from '../shared/types'
 import { DetailedProgrammeRights } from './detailedProgrammeRights'
 import { Language } from './language'
-import { Role } from './role'
 
 export type ExpandedUser = InferAttributes<User> & {
   iamGroups: string[]
