+++ services/shared/types/index.ts
@@ -1,2 +1,3 @@
 export type { Name } from './name'
 export type { Release } from './release'
+export type { Role } from './role'
