+++ services/shared/types/role.ts
@@ -1,0 +1,9 @@
+export type Role =
+  | 'admin'
+  | 'courseStatistics'
+  | 'facultyStatistics'
+  | 'fullSisuAccess'
+  | 'katselmusViewer'
+  | 'openUniSearch'
+  | 'studyGuidanceGroups'
+  | 'teachers'
